docs: 파일로 컴포넌트 분리 및 생성, PR 리뷰 내용 적용
