+++ C3-4T2D/View/ProjectView/Component/CommentModal.swift
@@ -1,0 +1,14 @@
+//
+//  CommentModal.swift
+//  C3-4T2D
+//
+//  Created by 차원준 on 5/30/25.
+//
+
+import SwiftUI
+
+struct CommentModal: View {
+    var body: some View {
+        Text("코멘트 창입니다.")
+    }
+}
